requirements/common.txt
beautifulsoup4==4.5.3
celery==4.0.2
Flask==0.12
Flask-HTTPAuth==3.2.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.3
Flask-Moment==0.5.1
Flask-QRcode==0.10.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
flower==0.9.1